pager内のcomposeをContentとして切り出し
